+++ src/pages/GroupsPage.vue
@@ -1,11 +1,23 @@
 <template>
   <q-page class="column q-pa-md bg-grey-1">
-    <h4
-      class="q-ma-none q-mt-md q-mb-lg"
-      data-cy="page_groups_title"
-    >
-      {{ $t('GroupsPage.text.title') }}
-    </h4>
+    <div class="row justify-center items-center">
+      <h4
+        class="q-ma-none q-mt-md q-mb-lg q-mr-auto"
+        data-cy="page_groups_title"
+      >
+        {{ $t('GroupsPage.text.title') }}
+      </h4>
+      <q-btn
+        outline
+        no-caps
+        color="primary"
+        class="bg-white"
+        data-cy="groups_button_add"
+        :label="$t('GroupsPage.text.add')"
+        :icon="$t('GroupsPage.icon.add')"
+        @click="$router.push('/add-group')"
+      />
+    </div>
     <groups-table
       :groups="groups"
       :detach-action="false"